Quick fix for long method names.

+++ java/net/mikekohn/java_grinder/Math.java
@@ -21,9 +21,9 @@ abstract public class Math
   public static int max(int n0, int n1) { return 0; }
 
   /** Calculated sine based on a 512 degree circle. */
-  public static int sin512(int value) { return 0; }
+  public static float sin512(int value) { return 0; }
 
   /** Calculated cosine based on a 512 degree circle */
-  public static int cos512(int value) { return 0; }
+  public static float cos512(int value) { return 0; }
 }
